pyproject.toml
[build-system]
requires = ["setuptools>=68", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "auto-reviewer"
version = "0.1.0"
requires-python = ">=3.10"
description = "Automated multi-agent document review system"
authors = [
    {name = "Auto-Reviewer", email = "noreply@auto-reviewer.local"}
]
dependencies = [
    "pydantic>=2,<3",
    "click>=8.1,<9",
    "rich>=13,<14",
    "python-docx>=1.0,<2",
    "pypdf>=4.0.1,<5.0.0",
    "sentence-transformers>=3,<4",
    "faiss-cpu>=1.8,<2",
    "chromadb>=0.4.24,<0.5.0",
    "numpy>=1.26,<3",
    "crewai>=0.28,<0.33",
    "crewai-tools>=0.4,<0.6",
    "httpx>=0.27,<0.28",
    "tiktoken>=0.7,<0.8",
]

[project.scripts]
auto-reviewer = "auto_reviewer.cli:main"

[tool.setuptools.packages.find]
where = ["src"]
include = ["auto_reviewer*"]